requirements.txt
counterfit
counterfit-shims-seeed-python-dht
paho-mqtt
pandas
